feat: add pagination support and class-validator integration for post management

+++ src/modules/posts/posts.service.ts
@@ -1,9 +1,15 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  ForbiddenException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { CreatePostDto } from './dto/create-post.dto';
 import { UpdatePostDto } from './dto/update-post.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Post } from './entities/post.entity';
 import { Repository } from 'typeorm';
+import { PaginationDto } from 'src/utils/pagination/pagination.dto';
+import { paginateAndFormat } from 'src/utils/pagination/pagination.util';
 @Injectable()
 export class PostsService {
   constructor(
@@ -11,7 +17,7 @@ export class PostsService {
     private readonly repo: Repository<Post>,
   ) {}
 
-  async createPostSync(
+  async createPost(
     createPostDto: CreatePostDto,
     employeeInfo: any,
   ): Promise<Post> {
@@ -25,39 +31,72 @@ export class PostsService {
     return post;
   }
 
-  async savePost(postData: any, employee: any) {
-    const post = this.repo.create({
-      ...postData,
-      employeeId: employee.employeeId,
-      employeeFullName: employee.fullName,
+  async getPostList(dto: PaginationDto) {
+    const { page = 1, pageSize = 10 } = dto;
+
+    const query = this.repo.createQueryBuilder('post')
+      .leftJoinAndSelect('post.reactionCounts', 'reactionCounts')
+      .orderBy('post.createDate', 'DESC');
+
+    return paginateAndFormat(query, {
+      page: Number(page),
+      pageSize: Number(pageSize),
+      useQueryBuilder: true,
+      queryBuilder: query,
     });
-    return this.repo.save(post);
   }
 
   async findAll(): Promise<Post[]> {
-    return await this.repo.find();
+    return await this.repo.find({
+      order: { createDate: 'DESC' },
+      relations: ['reactionCounts'],
+    });
   }
 
   async findOne(id: string): Promise<Post> {
-    const post = await this.repo.findOne({ where: { id } });
+    const post = await this.repo.findOne({
+      where: { id },
+      relations: ['postComments', 'reactionCounts'],
+    });
     if (!post) {
       throw new NotFoundException('This Post not found');
     }
     return post;
   }
 
-  async update(id: string, updatePostDto: UpdatePostDto): Promise<boolean> {
+  async update(
+    id: string,
+    updatePostDto: UpdatePostDto,
+    employeeId: string,
+  ): Promise<boolean> {
+    const post = await this.repo.findOne({ where: { id } });
+    if (!post) {
+      throw new NotFoundException('This Post not found');
+    }
+
+    if (post.employeeId !== employeeId) {
+      throw new ForbiddenException('You can only update your own posts');
+    }
+
     const updatePost = await this.repo.update(id, updatePostDto);
     if (updatePost.affected === 0) {
-      throw new NotFoundException('This Post not found');
+      throw new NotFoundException('Failed to update post');
     }
     return true;
   }
 
-  async delete(id: string): Promise<boolean> {
+  async delete(id: string, employeeId: string): Promise<boolean> {
+    const post = await this.repo.findOne({ where: { id } });
+    if (!post) {
+      throw new NotFoundException('This Post not found');
+    }
+
+    if (post.employeeId !== employeeId) {
+      throw new ForbiddenException('You can only update your own posts');
+    }
     const deletePost = await this.repo.delete(id);
     if (deletePost.affected === 0) {
-      throw new NotFoundException('This Post not found');
+      throw new NotFoundException('Failed to delete post');
     }
     return true;
   }
